Make gutter resizeable, Change grouping to commits, Pull first commit date
Added event listeners and logic to allow resizing the gutter by clicking
and dragging on the edge.

Added a method to the git helper to fetch the first commit date.

Added the date and author name to the gutter for each grouped commit.

Changed the colour idicator to show as a border instead of a background.