Improve logic to check whether Layer is installed
Instead of looking for Layer in /Applications, we now look for the
Layer Application Support directory in case the app has not been moved
to /Applications (e.g. kept in ~/Downloads).

Also, we now check for the existence of Layer at every boot of the
plugin so users who used the plugin before downloading Layer just have
to reboot Atom once they get Layer to activate the communication.